Allow setting a different file search class
This is made possible by extracting AbstractFileSystemResolver
