+++ lib/riiif.rb
@@ -4,6 +4,7 @@ require 'riiif/engine'
 module Riiif
   extend ActiveSupport::Autoload
   autoload :Image
+  autoload :AbstractFileSystemResolver
   autoload :FileSystemFileResolver
   autoload :HTTPFileResolver
   autoload :Routes
